refactor : search.tsx 리팩토링 search page 헤더 , 필터 container 컴포넌트로 분리해서 리팩토링 완료
